# packages/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FEVO Event Feed API
  version: 1.0.0
  description: |
    REST API for the FEVO Event Feed system. Provides a public feed of events/offers
    for embeddable widgets, plus admin endpoints for managing offers, organizations,
    segments, API keys, and FEVO integration.

    ## Authentication

    **Public endpoints** require an API key via the `x-api-key` header.
    Keys are issued through the admin API and hashed (SHA-256) before storage.

    **Admin endpoints** require an internal auth token via the `x-internal-auth` header.

    **Webhook endpoints** require the FEVO webhook secret via `x-fevo-webhook-secret`.

    ## Rate Limiting

    - Public feed: **100 req/min** per API key (configurable per key)
    - Admin endpoints: **30 req/min**
    - Rate limit headers are included in all responses

  contact:
    name: FEVO Engineering
  license:
    name: MIT
servers:
  - url: https://fevo-event-feed-api.onrender.com
    description: Production (Render)
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v1/event-feed/admin/offers/stats:
    get:
      tags: [Admin - Offers]
      summary: Get offer statistics
      description: Counts by status, distribution state, and organization.
      security:
        - InternalAuth: []
      responses:
        "200":
          description: Offer stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                      active:
                        type: integer
                      inactive:
                        type: integer
                      sold_out:
                        type: integer
                      deleted:
                        type: integer
                      distribution_enabled:
                        type: integer
                      distribution_disabled:
                        type: integer
                      by_organization:
                        type: array
                        items:
                          type: object
                          properties:
                            organization_id:
                              type: string
                              nullable: true
                            organization_name:
                              type: string
                              nullable: true
                            count:
                              type: integer
  /api/v1/event-feed/admin/offers:
    get:
      tags: [Admin - Offers]
      summary: List offers (admin)
      description: Paginated list with search, sort, and status filters.
      security:
        - InternalAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, sold_out, deleted]
        - name: distribution_enabled
          in: query
          schema:
            type: boolean
        - name: organization_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Searches title, venue name, venue city, org name
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, date, title, price_min, status]
            default: created_at
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated offers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Offer"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
  /api/v1/event-feed/admin/offers/{offerId}:
    get:
      tags: [Admin - Offers]
      summary: Get single offer
      security:
        - InternalAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Offer"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Admin - Offers]
      summary: Update offer (enrichment)
      description: |
        Enrich or edit offer fields. Only provided fields are updated.
        Useful for adding editorial content, correcting data, or overriding
        synced values.
      security:
        - InternalAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Offer"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────────────────────────────────
  # ADMIN — DISTRIBUTION
  # ──────────────────────────────────────────────────────────────────────────
  /api/v1/event-feed/admin/offers/{offerId}/distribution:
    get:
      tags: [Admin - Distribution]
      summary: Get distribution status
      security:
        - InternalAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Distribution status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DistributionStatus"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Admin - Distribution]
      summary: Toggle distribution
      description: |
        Enable or disable an offer in the public feed. Only active offers
        can be distribution-enabled.
      security:
        - InternalAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated distribution status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DistributionStatus"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Cannot enable distribution for inactive offer

  # ──────────────────────────────────────────────────────────────────────────
  # ADMIN — KILL SWITCH
  # ──────────────────────────────────────────────────────────────────────────
  /api/v1/event-feed/admin/kills:
    get:
      tags: [Admin - Kill Switch]
      summary: List active kills
      security:
        - InternalAuth: []
      responses:
        "200":
          description: Active kill records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Kill"
  /api/v1/event-feed/admin/kills/offer:
    post:
      tags: [Admin - Kill Switch]
      summary: Kill a single offer
      description: Immediately removes an offer from the public feed.
      security:
        - InternalAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [offerId]
              properties:
                offerId:
                  type: string
                  minLength: 1
                reason:
                  type: string
      responses:
        "201":
          description: Kill created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Kill"
  /api/v1/event-feed/admin/kills/organization:
    post:
      tags: [Admin - Kill Switch]
      summary: Kill all offers for an organization
      security:
        - InternalAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orgId, reason]
              properties:
                orgId:
                  type: string
                  minLength: 1
                reason:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Kill created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Kill"
  /api/v1/event-feed/admin/kills/{killId}/restore:
    post:
      tags: [Admin - Kill Switch]
      summary: Restore a kill
      description: Re-enables the killed offer/organization in the feed.
      security:
        - InternalAuth: []
      parameters:
        - name: killId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Kill restored
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Kill"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────────────────────────────────
  # ADMIN — ORGANIZATIONS
  # ──────────────────────────────────────────────────────────────────────────
  /api/v1/event-feed/admin/organizations:
    get:
      tags: [Admin - Organizations]
      summary: List organizations
      security:
        - InternalAuth: []
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
  /api/v1/event-feed/admin/organizations/{orgId}:
    get:
      tags: [Admin - Organizations]
      summary: Get organization
      security:
        - InternalAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────────────────────────────────
  # ADMIN — SEGMENTS
  # ──────────────────────────────────────────────────────────────────────────
  /api/v1/event-feed/admin/segments:
    post:
      tags: [Admin - Segments]
      summary: Create segment
      security:
        - InternalAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, type]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                slug:
                  type: string
                  pattern: "^[a-z0-9-]+$"
                  minLength: 1
                  maxLength: 255
                type:
                  type: string
                  enum: [theme, geography, partner, custom]
                rules:
                  type: object
                  additionalProperties: true
                is_curated:
                  type: boolean
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Segment"
        "409":
          description: Slug already exists
...
  /api/v1/event-feed/admin/api-keys/{keyId}/revoke:
    post:
      tags: [Admin - API Keys]
      summary: Revoke API key
      security:
        - InternalAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ApiKeyInfo"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Partner API key (format `efeed_XXXX`)
    InternalAuth:
      type: apiKey
      in: header
      name: x-internal-auth
      description: Internal admin auth token
    WebhookSecret:
      type: apiKey
      in: header
      name: x-fevo-webhook-secret
      description: FEVO webhook shared secret
...